fitting-service/requirements.txt
absl-py==1.1.0
addict==2.4.0
aiohttp==3.8.1
aiosignal==1.2.0
albumentations==1.2.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
autocfg==0.0.8
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cloths-segmentation==0.0.2
cPython==0.0.6
cycler==0.11.0
Cython==0.29.30
Django==4.0.6
djangorestframework==3.13.1
dnspython==2.2.1
docker-pycreds==0.4.0
efficientnet-pytorch==0.6.3
fonttools==4.34.4
frozenlist==1.3.0
fsspec==2022.5.0
gitdb==4.0.9
GitPython==3.1.27
gluoncv==0.10.5.post0
google-auth==2.9.1
google-auth-oauthlib==0.4.6
graphviz==0.8.4
grpcio==1.47.0
idna==3.3
ifaddr==0.2.0
iglovikov-helper-functions==0.0.53
imagecorruptions==1.1.2
imageio==2.19.3
importlib-metadata==4.12.0
iniconfig==1.1.1
joblib==1.1.0
jpeg4py==0.1.4
kiwisolver==1.4.4
lxml==4.9.1
Markdown==3.3.7
matplotlib==3.5.2
multidict==6.0.2
munch==2.5.0
mxnet==1.6.0
networkx==2.8.4
ninja==1.10.2.3
numpy==1.23.1
oauthlib==3.2.0
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
packaging==21.3
pandas==1.4.3
pathtools==0.1.2
Pillow==9.2.0
pluggy==1.0.0
portalocker==2.5.1
pretrainedmodels==0.7.4
promise==2.3
protobuf==3.19.4
psutil==5.9.1
py==1.11.0
py-cpuinfo==8.0.0
py-eureka-client==0.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyDeprecate==0.3.2
pymongo==4.1.1
pyparsing==3.0.9
pytest==7.1.2
pytest-benchmark==3.4.1
python-dateutil==2.8.2
pytorch-lightning==1.6.5
pytorch-toolbelt==0.5.1
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
qudida==0.0.4
requests==2.28.1
requests-oauthlib==1.3.1
reserve==0.3
rsa==4.8
scikit-image==0.19.3
scikit-learn==1.1.1
scipy==1.8.1
segmentation-models-pytorch==0.2.1
sentry-sdk==1.7.1
setproctitle==1.2.3
shortuuid==1.0.9
six==1.16.0
smmap==5.0.0
sqlparse==0.4.2
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
threadpoolctl==3.1.0
tifffile==2022.5.4
timm==0.4.12
tomli==2.0.1
torch==1.12.0
torchmetrics==0.9.2
torchvision==0.13.0
tqdm==4.64.0
typing_extensions==4.3.0
urllib3==1.26.10
wandb==0.12.21
Werkzeug==2.1.2
whitenoise==6.2.0
yacs==0.1.8
yarl==1.7.2
zipp==3.8.1
